loader: comparison between signed and unsigned integer expressions
/code/illumos-gate/usr/src/boot/lib/libstand/rarp.c:160:19: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
  if (n == -1 || n < sizeof(struct ether_arp)) {
                   ^
cc1: all warnings being treated as errors

+++ usr/src/boot/lib/libstand/rarp.c
@@ -157,7 +157,7 @@ rarprecv(struct iodesc *d, void **pkt, void **payload, time_t tleft,
 
 	n = readether(d, &ptr, (void **)&ap, tleft, &etype);
 	errno = 0;	/* XXX */
-	if (n == -1 || n < sizeof(struct ether_arp)) {
+	if (n == -1 || n < (ssize_t)sizeof(struct ether_arp)) {
 #ifdef RARP_DEBUG
 		if (debug)
 			printf("bad len=%d\n", n);
